refactor get_search_regex
