+++ bodo/transforms/distributed_analysis.py
@@ -1760,7 +1760,8 @@ class DistributedAnalysis:
                 )
 
             return
-        if fdef == ("get_search_regex", "bodo.libs.array"):
+
+        if fdef == ("get_search_regex", "bodo.hiframes.series_str_impl"):
             # out_arr and in_arr should have the same distribution
             new_dist = _meet_array_dists(
                 self.typemap, rhs.args[0].name, rhs.args[3].name, array_dists
